src/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
google-genai
python-dotenv
requests
